auth/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.0.1
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.112.2
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.8.2
pydantic-settings==2.10.1
pydantic_core==2.20.1
python-dotenv==1.0.1
python-jose==3.5.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
